requirements.txt
Django==4.0.4
djangorestframework==3.13.1
django-cors-headers==3.12.0
django-extensions==3.1.5
redis==4.3.1
psycopg2-binary==2.9.3
django-environ==0.8.1
celery==5.2.6
beautifulsoup4==4.11.1
requests==2.27.1
Pillow==9.1.1
django-filter==21.1
